+++ src/LaravelPausableBatchClass.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Digiloop\LaravelPausableBatch;
-
-class LaravelPausableBatchClass
-{
-}